Push filler token to prevent empty tokens array
Previously empty lines in clojure files were generating empty
token arrays since the parser we bailing when at the end of the line
with no tokens and the rule stack wasn't increasing.

This case also needed to push a filler token for the empty line to prevent
a line with no tokens.