Break when loop detected and stack can't be popped
Previously the stack would be popped even when it contained a single rule
causing the next loop to try to index into an empty rule stack.

Now if a loop is detected with a single rule in the stack the loop
is broken out of and a filler token is added for the remaining positions
in the line.

Closes #11