:bug: Fix wrong scopes applied to lines following a very long line.
Make sure the rule stack is copied, so that modifying it does not modify the
original. This was causing the stack to be 'restored' to the wrong state when
the max tokens per line limit was reached and there were no initial rules.

See https://github.com/atom/atom/issues/1667#issuecomment-69776121