Store scopes and tokens as integers and strings in a flat array
Instead of token objects, we now represent tokenized content as a
sequence of scope begin/end tags interleaved with strings. The scope
tags are represented by integer ids into a global symbol table on the
grammar registry. We indicate an end tag with the same integer id as
the start tag, but make it negative.

This should represent a big memory savings over having an object with
an array of scopes for every single token.