Represent line scopes as an array of integers
Previously, I was representing the content of a line as a mix of strings
and integers. This commit replaces the strings with positive integers
representing the length of the substring at that location. Scope tags
are switched to negative integers to distinguish them from tokens, with
odd integers representing scope begin tags and even integers
representing end tags.