Don’t report token boundaries beyond the end of lines
We append `\n` to the end of lines before parsing. Some patterns match
against this appended character, which was causing us to return an array
of tags that exceed the length of the original line. Now we omit any
positive tags that exceed the length of the original line, but continue
to include the scope open/close tags that occurred there.

Signed-off-by: Katrina Uychaco <kuychaco@github.com>